More work on Deleting Players from Team
When deleting players from a team needs to be added that when cancel is
pressed changes aren't pernamanently applied. Will have to have a list
of players to delete and it will perform the delete once Done is pressed

+++ JulyTest/LeaderBoardApp/ModalControl/PlayerSelectDelete.cs
@@ -24,8 +24,12 @@ namespace LeaderBoardApp.ModalControl
             var playerSelected = modalSelect.GetSelectedPlayer();
             if (playerSelected != null)
             {
-                var playerID = (int)playerSelected;
-                DeletePlayer(playerID);
+                 var result = MessageBox.Show("Are you sure you want to Delete?", "Confirmation", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                 if (result == MessageBoxResult.Yes)
+                 {
+                     var playerID = (int)playerSelected;
+                     DeletePlayer(playerID);
+                 }
             }
         }
